# internal/surface/http/openapi.json
{
  "components": {
    "parameters": {
      "Id": {
        "description": "The resource's ID, or (for an account or category) its unique name.",
        "in": "path",
        "name": "id",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "InvalidInput": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        },
        "description": "The request failed validation."
      },
      "NotFound": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        },
        "description": "No such resource."
      },
...
    },
    "schemas": {
      "Account": {
        "properties": {
          "archived": {
            "type": "boolean"
          },
          "archived_at": {
            "description": "Set only once the account is archived.",
            "format": "date",
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "institution": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "opening_balance": {
            "description": "A plain decimal amount. Always a JSON string, in the sibling currency field's currency - never a number.",
            "format": "money",
            "type": "string"
          },
          "opening_balance_date": {
            "description": "The date the opening balance is stated as of.",
            "format": "date",
            "type": "string"
          },
          "sort_order": {
            "description": "Where the account sorts in a list; lower comes first.",
            "type": "integer"
          },
          "type": {
            "enum": [
              "bank",
              "cash",
              "credit_card",
              "wallet",
              "investment",
              "loan",
              "other"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "currency",
          "opening_balance",
          "sort_order",
          "archived"
        ],
        "type": "object"
      },
      "AccountEnvelope": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Account"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "AccountListEnvelope": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Account"
            },
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "ApiToken": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "expires_at": {
            "description": "Absent for a token that doesn't expire.",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "last_used_at": {
            "description": "Absent if the token has never been used.",
            "format": "date-time",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "revoked_at": {
            "description": "Absent for a token that's still live.",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "created_at"
        ],
        "type": "object"
      },
      "ApiTokenListEnvelope": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ApiToken"
            },
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
      "CreateAPITokenRequest": {
        "properties": {
          "expires_at": {
            "description": "Omit for a token that never expires.",
            "format": "date",
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "CreateAPITokenResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "expires_at": {
            "description": "Absent for a token that doesn't expire.",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "token": {
            "description": "The plaintext token. Shown only this once - store it now; it can't be retrieved again.",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "created_at",
          "token"
        ],
        "type": "object"
      },
      "CreateAPITokenResponseEnvelope": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CreateAPITokenResponse"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "CreateAccountRequest": {
        "properties": {
          "currency": {
            "type": "string"
          },
          "institution": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "opening_balance": {
            "description": "Defaults to zero.",
            "format": "money",
            "type": "string"
          },
          "opening_balance_date": {
            "description": "The date the opening balance is stated as of. Omit to state it as of today.",
            "format": "date",
            "type": "string"
          },
          "sort_order": {
            "description": "Where the account sorts in a list; lower comes first.",
            "type": "integer"
          },
          "type": {
            "enum": [
              "bank",
              "cash",
              "credit_card",
              "wallet",
              "investment",
              "loan",
              "other"
            ],
            "type": "string"
          }
        },
        "required": [
          "name",
          "type"
        ],
        "type": "object"
      },
...
      "Error": {
        "properties": {
          "code": {
            "enum": [
              "conflict",
              "internal",
              "invalid_input",
              "not_allowed",
              "not_found",
              "precondition_failed",
              "unauthenticated",
              "unavailable"
            ],
            "type": "string"
          },
          "details": {
            "additionalProperties": {},
            "description": "Safe, structured extra context - candidate names, valid options.",
            "type": "object"
          },
          "field": {
            "description": "The request field or CLI flag this error attaches to, when there is one.",
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "Ok": {
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "type": "object"
      },
      "OkEnvelope": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Ok"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "PatchAccountRequest": {
        "properties": {
          "name": {
            "description": "The account's new name.",
            "nullable": true,
            "type": "string"
          },
          "opening_balance": {
            "description": "The opening balance to re-declare.",
            "format": "money",
            "nullable": true,
            "type": "string"
          },
          "opening_balance_date": {
            "description": "The date the new opening balance is stated as of. Only read alongside opening_balance.",
            "format": "date",
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "bodger's read/write REST API. This is the web UI's only interface (milestone 2 generates its TypeScript client from this description) and is available to external clients and integrations. Every route but /healthz and POST /api/v1/auth/login requires a session cookie (from login) or an Authorization: Bearer API token; a cookie-authenticated state-changing request also needs the X-Bodger-CSRF header. The server binds a loopback address only unless a password has been set. Monetary values are always JSON strings, never numbers.",
    "title": "bodger REST API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/v1/accounts": {
      "get": {
        "operationId": "listAccounts",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountListEnvelope"
                }
              }
            },
            "description": "Every account, in name order."
          }
        },
        "summary": "List every account."
      },
      "post": {
        "operationId": "createAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountEnvelope"
                }
              }
            },
            "description": "The created account."
          },
          "422": {
            "$ref": "#/components/responses/InvalidInput"
          }
        },
        "summary": "Create an account."
      }
    },
    "/api/v1/accounts/{id}": {
      "delete": {
        "description": "This hides the account from pickers; its history remains. There is no hard delete for an account with any history.",
        "operationId": "archiveAccount",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountEnvelope"
                }
              }
            },
            "description": "The archived account."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "Archive an account."
      },
      "get": {
        "operationId": "getAccount",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountEnvelope"
                }
              }
            },
            "description": "The account."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "Fetch one account by ID or unique name."
      },
      "patch": {
        "description": "The request body must set exactly one of \"name\" or \"opening_balance\" - these are two separate operations at the application layer, and a request naming both or neither is rejected.",
        "operationId": "patchAccount",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountEnvelope"
                }
              }
            },
            "description": "The updated account."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/InvalidInput"
          }
        },
        "summary": "Rename an account, or re-declare its opening balance."
      }
    },
...
    "/api/v1/auth/tokens": {
      "get": {
        "description": "Includes revoked and expired tokens, for history - not only what's currently live.",
        "operationId": "listAPITokens",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiTokenListEnvelope"
                }
              }
            },
            "description": "Every API token, newest first."
          }
        },
        "summary": "List every API token the acting user owns."
      },
      "post": {
        "description": "The plaintext token is returned exactly once, in this response - it can't be retrieved again.",
        "operationId": "createAPIToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAPITokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAPITokenResponseEnvelope"
                }
              }
            },
            "description": "The created token, including its one-time plaintext value."
          },
          "422": {
            "$ref": "#/components/responses/InvalidInput"
          }
        },
        "summary": "Issue a new API token."
      }
    },
    "/api/v1/auth/tokens/{id}": {
      "delete": {
        "description": "The token stops authenticating immediately, but remains listed (soft revocation).",
        "operationId": "revokeAPIToken",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkEnvelope"
                }
              }
            },
            "description": "The token was revoked."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "Revoke an API token."
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}